qwen/requirements_qwen.txt
torch
transformers
peft
accelerate
huggingface_hub
fastapi
uvicorn
pydantic
python-dotenv
